Add `~should_print_summary` to `Profiling_js.with_profiling`
Summary:
`Profiling_js.with_profiling` now has a new boolean option. When true,
it will print out a table of all the profiling events. This table is a useful
summary of where all the wall and cpu time went. It also shows you if we're
neglecting to profile some expensive code.

Reviewed By: samwgoldman

Differential Revision: D6256198

fbshipit-source-id: de7e5abe450d78680419f5fe4ba9609bfc275f85