FlowServerMonitorServer
Summary:
This awkwardly named module is the part of the Flow server
monitor which manages the Flow server. It's responsible for starting up
a server instance, noticing when it dies, and either exiting or
restarting the server.

Other threads who want to send messages to the server instead send the
messages to `FlowServerMonitorServer`. It will queue them up in a
stream. The queued up messages will be sent to the server once a server
instance is running.

Sometimes requests are sent to the server, but the server dies before it
can respond. In that case, we re-queue the requests.

Reviewed By: mroch

Differential Revision: D6266103

fbshipit-source-id: f850456a8c4316159bea6c460b34af7c7dd27ae6