$Call and friends bypasses intersection prep
Summary:
Calling intersections (of functions) in general needs some preparation, since the arguments might be `OpenT`s which should be concretized before `try_insersection` is invoked.

However, with $Call we know that arguments will be annotations, so they don't need to be concretized.

Note that as part of concretization, unions in arguments would be split, which in general increases the likelihood of matching cases of the intersection. However with $Call this can is problematic unless the corresponding returns are joined. Avoiding concretization altogether avoids this problem; the slight reduction in precision could be recovered in later diffs.

Reviewed By: calebmer

Differential Revision: D6258254

fbshipit-source-id: 9248ee55417599c91fc4947721cd5ab3971c245c