connect_and_make_request
Summary:
Many of our client commands connect to the server, send a
request, and then read a response. This combines those 3 common steps
into a single helper function.

A future diff will replace the implementation of
`connect_and_make_request` with an implementation that talks to the Flow
server monitor.

Reviewed By: mroch

Differential Revision: D6272550

fbshipit-source-id: 5cdf00161d749e66776f0074b29260900b5ad798