fix Object.keys on a dictionary of string enums
Summary:
previously, `GetKeysT` (which powers `Object.keys`) on a dictionary
ignored the type of the dictionary keys and always returned a generic `StrT`,
because keys are always strings.

this diff introduces a `ToStringT` use type which passes through strings
unchanged, or returns a `StrT` otherwise. this is a loose approximation of the
[`ToString`](https://tc39.github.io/ecma262/#sec-tostring) abstract operation in the spec
which we can extend later to properly convert other types to the appropriate
string literals, e.g. `null` to a `StrT ("null")` or `NumT`s via ocaml-dtoa.

in this way, Object.keys on a dictionary of a string enum will return the
string literals like you'd expect.

Reviewed By: avikchaudhuri

Differential Revision: D6233861

fbshipit-source-id: b1151ae1a32488531d6e08fa91f2554e177095c0