Improve top level reason description and remove description qualifications
Summary:
In the new error message format I’m working on, we show top level reason descriptions to the user fairly frequently. Top level reason descriptions are the description for the tvar or annotation before it gets resolved. Traditionally we mostly show the reason descriptions after types have been resolved. `string` vs variable `x`. There are a handful of clowny top level reason descriptions that I change in this diff.

In this diff I also remove a lot of qualifications like “type application:”, “class:”, or “identifier `x`”. These qualifications reflect our internal representation of a type, not necessarily the representation a user thinks of. When we generate error messages as a sentence and not a collection of sentence fragments, these qualifications don’t read well.

In addition to being good for a future error format, these changes also make errors easier to parse in our current error format.

This diff used to be deeper in my error messages stack. I moved it to the top of my stack and squashed a couple of other commits into it. If you would like to see the difference individual changes in this diff make, let me know and I can “unsquash.”

Reviewed By: mroch

Differential Revision: D6055345

fbshipit-source-id: 388d6ac3e50722a48024f8d833996acd8e12f21d