Add MonitorRPC.status_update calls
Summary:
Since `MonitorRPC` is disabled, these are no-ops. But once it's
enabled, these will send the current server status to the monitor.

The multiworker progress stuff is pretty primitive. It only sends
progress updates when `next ()` is called to get more work, so we don't
**really** know how many jobs have finished. But this is good enough to
give our users a rough idea of the status, and we can make it smarter in
the future, like what mroch built in {D6215173}.

Reviewed By: samwgoldman

Differential Revision: D6272819

fbshipit-source-id: c0ffa1a35a9f7aba046fbe09b5ee15e5f1be8b86