Use XXH_STATIC_LINKING_ONLY
Summary:
The state struct for xxHash is incomplete, which means it's not possible to
stack allocate the struct. Instead, you have to use XXH64_createState to get a
pointer to the heap.

If you define XXH_STATIC_LINKING_ONLY, though, the struct definition is
available. This is done to protect dynamically linked xxHash users from API
changes. Since we statically link xxHash, we don't need this protection.

By avoiding the heap allocation, we can just copy the struct directly into the
ocaml custom val, so everything will get collected during GC. That is, we don't
need to install a finalizer that frees the struct.

Unlike state updates, which are very hot, this code only runs once per sig
context during merge -- it's not very hot. Still, this was bugging me.

Reviewed By: mroch

Differential Revision: D6253913

fbshipit-source-id: 15ede7e27048d5030fc2345f9cb1563ff201f3b9