fix non-termination for predicates on classes
Summary:
A class value `ThisClass(i_with_this)` when used as a type `Class(i)` is "fixed"
with the equation `i = i_with_this { this := i }`.

The tvar created to represent `i` might be generated over and over again,
leading to non-termination.

This diff fixes (hah) by the use of a cache.

Reviewed By: samwgoldman

Differential Revision: D6293902

fbshipit-source-id: 4ccd504918f9cbd882fde433c19775c02ddb1ba9