Don't bother canonicalizing xxHash digests
Summary:
Since the digest lives on a single flow server on a single host, we don't need
to canonicalize (convert to network byte order).

Unlike hash state updates, this is not particularly hot code, happening once
per sig context per merge. Still, we don't hash ints in a portable way, so this
code is potentially misleading.

Reviewed By: mroch

Differential Revision: D6254250

fbshipit-source-id: 17ab730cf40d0f490cbc32d4aea0ce35027e2538