Fix scrollbar color
Body's background color affects scrollbar color in certain operating systems (macOS, iOS…). In this case the background was dark, so the OS flips the scrollbar color to white, which is almost invisible on website's mostly light background.

This change brings back the dark scrollbar, which is completely visible on darker parts of the website.