[PR] Fix scrollbar color
Summary:
Body's background color affects scrollbar color in certain operating systems (macOS, iOS…). In this case the background was dark, so the OS flips the scrollbar color to white, which is almost invisible on website's mostly light background.

This change brings back the dark scrollbar, which is completely visible on darker parts of the website.

<img width="285" alt="screen shot 2017-11-12 at 20 41 21" src="https://user-images.githubusercontent.com/471278/32702621-1ccd2a78-c7ea-11e7-8fe7-0143b5f94fa8.png">

<img width="289" alt="screen shot 2017-11-12 at 20 41 33" src="https://user-images.githubusercontent.com/471278/32702623-1fdf4034-c7ea-11e7-90f0-919f9627a11a.png">
Closes https://github.com/facebook/flow/pull/5326

Differential Revision: D6310673

Pulled By: mroch

fbshipit-source-id: 277f924943272f0b0a6e01ce7ba6f5b09b94ec2d