[PR] lib/BOM: add missing "speed" to Coordinates
Summary:
According to the [W3C specification](https://dev.w3.org/geo/api/spec-source.html#coordinates_interface), there is a "speed" property on Coordinates objects.

This PR merely adds this to the BOM class declarations in FlowType, as it was missing.

I have screenshots from Safari and Chrome that show that this part of the specification is (sort of) implemented in browsers:

![chrome-coordinates-speed](https://cloud.githubusercontent.com/assets/479816/22008872/2f0c14cc-dcd3-11e6-86b6-7fa2eb97e26f.png)
<img width="560" alt="safari-coordinates-speed" src="https://cloud.githubusercontent.com/assets/479816/22008873/2f4c94e8-dcd3-11e6-8b13-c69f271cef6d.png">
Closes https://github.com/facebook/flow/pull/3198

Differential Revision: D4861103

Pulled By: mroch

fbshipit-source-id: 62fc17a123726fd473261c4e9f87b90619dd7dcf