Remove require_loc from resolved requires heap
Summary:
We can recover this information from the file sig heap. It does not need to be
duplicated here.

This diff also tightens up the data sent over to the client. We used to send a
set of module names and a map from module name (converted to strings) to
require locations.

On the client, locations form the map are find_unsafe'ed with values from the
set (again, converted to strings).

Instead just send a map from module names to require locations. No string
conversion required.

Reviewed By: gabelevi

Differential Revision: D6262101

fbshipit-source-id: 928d926d1c59446acedee88a436bf4a45cdff3cb