Add parse error for missing commas in export specifiers
Summary:
Currently we parse `export { foo bar }` as if there were a comma between foo
and bar, but this should be a parse error. Trailing commas are allowed, same as
import specifiers.

Facbeook:
We don't recover very well for `declare export type`, and this exacerbates
that, but I don't want to fix it right now. I think it's fine.

Reviewed By: mroch

Differential Revision: D6320699

fbshipit-source-id: 053c6016dcbffe917249091ccf86657c38bc10ff