runtests.sh improvements
Summary:
adds an `-l` flag to list all tests, and a `-t TEST_NAME` flag that
runs a specific test. also adds a `-q` flag to quiet down the output, and a `-d DIR`
flag to run the tests in `DIR/tests/` instead of the ones that are next to
runtests.sh.

since `getopts` doesn't parse options after the first positional arg, I also
added optional `-b` and `-f` flags for the binary name and filters. you can
continue to do `runtests.sh bin/flow classes`, or you can now do
`runtests.sh -b bin/flow -f classes` so that you can add `-r` to the end of the
command instead of as the first arg.

Reviewed By: gabelevi

Differential Revision: D6306731

fbshipit-source-id: bb940cb7d49aa8b6bcbcbe08ca5deffedf26116f