Remove uses of `T` in input position of `$ReadOnlyArray`
Summary:
Consider:

```
declare class $ReadOnlyArray<+T> {
  [index: number]: T;
  includes(value: T): boolean;
}
```

In `includes()`, `T` is used in an input position, but `T` is positive which means `T` may only be used in output positions.

We incorrectly allow this in `declare class`, but disallow it in proper class declarations:

```
class $ReadOnlyArray<+T> {
  include(value: T): boolean {
    return false;
  }
}
```

[try-Flow](https://flow.org/try/#0MYGwhgzhAEAkBKBTMATA8gOxATwIICd8xsAeAagBUA+aAbwChpoBLDUAVxUQAoA3MEO0QAuaBQCUogEYB7GSGQY6jJtHyIALu3xKAZgIiIA3CoC+9c0A)

Users have also asked from time to time why they can't pass in arbitrary types to `includes()`. So this also improves the user experience.

Reviewed By: samwgoldman

Differential Revision: D6324909

fbshipit-source-id: 7ee070bcdc423f474b811659b82e4736a5a47c6f