Add more structure to export information in file sigs
Summary:
Before this change, a named export was simply an optional location, which
erases useful knowledge about the exports. For example, whether the export is a
namespace (e.g., `export * as x`) or is re-exported from another module (e.g.,
`export ...  from`). This diff adds the necessary structure to preserve this
information.

I also renamed `batch` to `star`, to help differentiate a namespace export,
which is named, from export *.

We can use this information to more precisely model dependencies. For example,
if module A depends on module B, and module B re-exports from module C, we can
chase the pointer from A to C directly before calculating the dependency graph.
My hope is that exploiting this information will help reduce cycles in the
current, coarser dep graph.

Reviewed By: nmote

Differential Revision: D6319755

fbshipit-source-id: c26964cd2b5b4836916ff8f8cd20d485f96cfe71