Combine ServerProt responses into single sum type
Summary:
The current code had a different type for each response, which
meant it was hard for the monitor to deal generically with any response.
The solution was to keep the response marshal'd and just pass around the
bytes.

This is a better and safer solution. Now we're less reliant on
marshaling the response.

Reviewed By: samwgoldman

Differential Revision: D6331119

fbshipit-source-id: eb9d6ed7abb592ad657afb4644a4d67c92092463