[ast] type param name is an identifier
Summary:
uses `Ast.Identifier.t` instead of `string` for type param names.

none of our tests changed, but presumably we'd be able to point at the `T` in `function f<T: U>()` now, instead of the whole `T: U`.

my motivation is to make sure that all identifiers are identifiers, not escaping as random `string`s. the loc tracking is a bonus.

Reviewed By: calebmer

Differential Revision: D6359595

fbshipit-source-id: 639cb0cccd531a2948dfc6fef58048c29f2d1b32