[hack] Informant passes through whether or not saved state is tiny
Summary:
Informant passes through whether or not saved state is tiny.

In the case where the informant specifies a saved state target, it
provides an everstore handle for the new server to download from.
What is lost is whether or not that handle belongs to a tiny saved state
or not. This funnels that through.

Reviewed By: jamesjwu

Differential Revision: D6364358

fbshipit-source-id: 425c5521dd779c1dc831de449314ef6ce8384530