[typing] allow getters and setters using string literals
Summary:
the error message shown when you do `get "foo"() {}` says that non-string-literal getters are not supported yet. but it is a string-literal getter...

the code looks like it's a refactoring accident, perhaps from when we made property values into variants. I came across this while changing the whole prop into a variant instead of just the value.

Reviewed By: samwgoldman

Differential Revision: D6367542

fbshipit-source-id: d8cdeb4923d892cbcbcf8939e58a332c14f113c8