Use requires from context in assume ground
Summary:
Before this change, a list of requires was passed to assume ground, which were
looked up in the context's require map to get tvars corresponding to imports.

Instead, we can just use the context's own list of import tvars. Since we need
to walk all imports, this is always reasonable to do.

Why do I want to do this? Well, I want to change the way we do requires.
Instead of each require corresponding to a string (the "source" of an
import/export from statement), we might pointer-chase through the dependencies,
so each named import needs to be a separate tvar.

For example, in `import { foo, bar } from "a"`, `foo` might point to module A,
but `bar` can point directly to module B, if A re-exports `bar` from B.

Stacked diffs will make the above clearer.

Reviewed By: avikchaudhuri

Differential Revision: D6383572

fbshipit-source-id: 89f96dd10d1e4a105e46670aa908317f74adbca1