Use all require/import locations
Summary:
Before this diff, when adding require tvars to a context, we only cared about
one of the locations of a reference. An implication of this is that we only
error about missing modules in a single location, even if the module is
required from multiple places.

The current behavior isn't too bad, but erroring on every require isn't the
only reason I want to do this. I'm planning to overhaul the way we connect
dependencies. Indexing requires by location instead of module reference is a
step toward this.

Reviewed By: avikchaudhuri

Differential Revision: D6383575

fbshipit-source-id: 0c3bab11caaf4c1a49289222d8c5acf1b6384240