Make declare modules less special-case-y
Summary:
Before this diff, processing a declare module statement required lots of
conditional checks.

For example, processing require/import statements had different behavior
depending on whether we were "in" a declare module or not.

After this diff, requires/imports and exports use the same code paths within a
declare module as without. The only difference is that `module_t_of_cx` returns
a different tvar, representing the exports of the declare module.

This approach is better because it makes it possible to re-write how
requires/imports work. A follow-up diff will make this clear, but my main goal
is to add tvars for individual named imports, rather than for files.

My main goal here is not to make declare module better, and indeed declare
module is still fairly confusing. I made the smallest change to unblock myself.
Further improvements in this space woud be very welcome, though.

Reviewed By: avikchaudhuri

Differential Revision: D6383576

fbshipit-source-id: 6b09023312a6bae27e59e4775b8f2788799bbd03