[PR] Improving the performance of test generation
Summary:
This PR improves the performance of test generation in two ways.
1. It calls Flow type checking programmatically, instead of invoking the flow process on command line.
2. It runs all the generated files through node in one shot, as opposed to invoking node repeatedly.
This improvement is *very* significant -- we can now generate and run 5000 test program in less than 5 seconds.  This means that we can explore larger state spaces.

To compile `testgen`, go to the root directory and type `make testgen`.
Closes https://github.com/facebook/flow/pull/5260

Reviewed By: samwgoldman

Differential Revision: D6235093

Pulled By: schandragit

fbshipit-source-id: 30c4c5ee32604846d08cd5631ec80765babb7f11