Always use identifier name as the top level reason description
Summary:
In the new error format I've been working on we use top level reason descriptions a lot more. For example the message might say "Cannot cast `x` to `y`" deriving `x` and `y` from the top level reason descriptions of types being cast in `statement.ml`.

I quickly discovered that the reason description was based on whether a variable was declared locally or whether that variable was imported. An imported variable is an `OpenT` with an `RIdentifier` reason description, but a local variable would be the underlying type. For example, a `NumT` with an `RNum` reason description.

This diff makes sure that all variables get an `RIdentifier` reason. We accomplish this by creating a new type variable for local variables, giving that a reason description of `RIdentifier`, and unifying it with the underlying type. This should not create a lot of overhead and makes a significant difference in error messages that become “Cannot cast `x` to `y`” instead of “Cannot cast intersection type to `y`”.

This diff also makes a significant difference for some of the reason descriptions improved in D6055345. Namely function call reason descriptions which get used a lot.

Reviewed By: avikchaudhuri

Differential Revision: D6267854

fbshipit-source-id: 7392c4a856d16ae3676440da9beace34abde49ef