Use the type alias name for the reason description of top level types
Summary:
"object type" is a generally unhelpful reason description. This diff changes the reason description for the top level type of an annotation to the name of that type. For example:

```
type O = {};
declare var o: ?O;
(o: O);
```

Now says "null is incompatible with `O`" instead of "null is incompatible with object type."

Another good example:

```
type O = {p: number};
({p: 'foo'}: O);
```

The top of the description now says “object literal. This type is incompatible with `O`” instead of “object literal. This type is incompatible with object type.”

This diff was inspired by motivations similar to those in D6267854. Unlike D6267854, however, we actually need to modify the reason description. In D6267854 we can settle for creating a noop tvar since we only care about the top level reason description. Here, however, we want the reason description after we’ve resolved tvars to have the alias name.

To accomplish this I introduce a `RTypeAlias of string * reason_desc` reason description and use `mod_reason_of_t` to update the reason description of the type returned by `Type_annotation.convert` in `statement.ml`. However, since we use reason descriptions in type checking, this strategy alone would be pretty dangerous. An example of some type checking that would break by changing this reason description is checking for `$Call` and `$ObjMap` when they take intersection functions. (See `DescFormat` for more examples.) So when `desc_of_reason` is called we implicitly unwrap `RTypeAlias` to the reason description we replaced. This preserves the expectation of code paths that depend on `desc_of_reason` for type-checking. In `Flow_error`, however, reason descriptions are purely for asthetics. So we alias `desc_of_reason` to a version that does not implicitly unwrap `RTypeAlias`.

Reviewed By: avikchaudhuri

Differential Revision: D6268601

fbshipit-source-id: 9306a02614f68bd0458512f7ed0ae2840b445597