Propagate use_op in exact object property not found
Summary:
In the following program, one error correctly has a `use_op`. The other does not. This diff fixes this.

In both cases there is an argument that both `use_op`s are redundant. This diff is focused on consistency. It may make sense to be consistent in the other direction.

```
declare var any: any;
((any: {||}): {|p: number|});
((any: {|p: number|}): {||});
```

[try-Flow](https://flow.org/try/#0CYUwxgNghgTiAEA3W8oDsCeAuVmDcAUABRHrbwDeAPlQL4CUO1ADjmgK4C2ARiDHfUIkyTKq3gcefAaIGEgA)

Reviewed By: avikchaudhuri

Differential Revision: D6355037

fbshipit-source-id: 69453cb313bd6ef86edba8d3c37243b3b0730dd7