Add use_op for class extension and implementation checks
Summary:
Consider:

```
class A {
  a: number;
}

class B extends A {
  a: string;
}
```

[try-Flow](https://flow.org/try/#0MYGwhgzhAECC0G8BQ1pgFzQHYFcC2ARgKYBOA3EgL5JKiQwBC0RAHgC5FYAmM8yqGaBDYkAllgDmFakA)

Currently, we get two errors for `number ~> string` and `string ~> number`. Both errors provide no context about what caused them and the `number ~> string` error is in a blatantly wrong location which we would never want to blame. (This poor location has delayed deploys in the past.)

This diff adds a `use_op` so that we put "`B` cannot extend `A`" above an incompatibility.

This leads to much better positioning. It is worth considering bumping up the field/method in the error message when that field/method is inside the class definition for better suppression locations. This should be addressed in the new error format I'm working on.

Reviewed By: avikchaudhuri

Differential Revision: D6327400

fbshipit-source-id: 63219c93aa9487645d06136a2f76592dbbfc6173