ignore result of getprop when speculating (like call)
Summary:
When checking a GetProp on an Intersection (of Objs), we should not involve the
result. This is important because the result might already be constrained
elsewhere (and involved in errors), and when involved in checking branches it
might cause the branches to error too early, causing misattribution of blame.

It is instructive to note that we follow a similar approach for Call already. We
should test all the other usual use types that carry results to confirm they
suffer from the same problem, and fix them as well.

Reviewed By: calebmer

Differential Revision: D6454063

fbshipit-source-id: 72b3f338364b17fad6a727a8fd225be13a21aee4