Don't refine opaque types based on their bounds
Summary:
If an opaque type is declared with a bound, values with that type can be
treated as a subtype of the declared bound. For example, given a type
definition `opaque type T: X = Y`, `T` is a subtype of `X`, but `X` is not a
subtype of `T`.

Now, say we refine a value with type `T` -- what should we get? If we get
anything other than `T`, we can't assign the refined value back to `T`. It's
important that refinements always produce a subtype of the thing being refined.
The only way to preserve this property is to never refine opaque types.

Reviewed By: avikchaudhuri

Differential Revision: D6430367

fbshipit-source-id: 4519976a324deccc5bb3ace371040821a83480a9