Fix declared opaque types with union-like bound
Summary:
We ensure that the underlying type is compatible with any declared bound. For
declared opaque types, there might not be an underlying type.

Before this change, we would check the opaque type itself against the bound,
which was an error.

This diff only performs the bound-is-supertype check if there is both a super
type and an underlying type.

Reviewed By: panagosg7

Differential Revision: D6432005

fbshipit-source-id: b38df21bf9e081b334e31fdb387f67f91942420e