Refactor function use_ops.
Summary:
The current function use_ops don't describe everything we need for descriptive error messages. This diff builds on top of the last which added a use_op slot to `CallT` to pass a `FunCall` use_op into `CallT` and build from there. This also adds a `FunCompatibility` use_op for knowing when two functions are being subtyped. This refactor and has very little impact on the test error messages.

Specifically, these refactors are all about allowing `FunParam` to track the location and position of individual function arguments in both calls and function subtyping. This allows me to create messages like: “Cannot cast `fn1` to `fn2` because in the second argument and then property `p` string is incompatible with number” while making sure that the error points to the second argument as the primary location and not string and number which could be anywhere. I demonstrate how this can be useful by changing the error message to state which parameter is incompatible. For instance, “the first parameter” or “the second parameter.”

It is also important that we have the `op` reason in `FunCall` for D6055682 where we want to position the error at the root location of the `use_op`.

The new `use_op` structure. I’m open to suggestions for changes:

- `FunCall`: When the function gets called, this is the root `use_op`.
- `FunCallMethod`: Same thing as `FunCall` but for method calls.
- `FunCompatibility`: When we do function subtyping we add this `use_op`. Similar to a `PropertyCompatibility` `use_op`.
- `FunParam`: We add this `use_op` in the multiflow code which handles both function calls and function subtyping. `FunParam` most commonly has a parent `use_op` of `FunCall`, `FunCallMethod`, or `FunCompatibility`.
- `FunReturn`: When we do function subtyping we use this and give it a parent of `FunCompatibility`.

Reviewed By: avikchaudhuri

Differential Revision: D6253791

fbshipit-source-id: facbca64cd3fb9a081142c75dc7ea0174e536c64