Always use the FunParam(FunCall) use_op location for the primary error location
Summary: Almost always, the `lower` reason for `FunParam(FunCall)` is the same as the `lower` reason we add to `FunParam` in the multiflow code. Except sometimes we miss a `reposition` or it is really hard to arrive at a correct `reposition` so we don't try. The `lower` in `FunParam(FunCall)` should always point to the right location.

Reviewed By: avikchaudhuri

Differential Revision: D6238474

fbshipit-source-id: 80402f1ac63cbc9fe603afa1d7722c3b41b207e1