Improve positioning for type destructor errors in implicitly instantiated PolyTs
Summary:
Consider:

```
declare function fn<O>(o: O): $PropertyType<O, 'p'>;
fn({x: 42});
```

[try-Flow](https://flow.org/try/#0CYUwxgNghgTiAEAzArgOzAFwJYHtVNQB4B5APgAocAueYgShoBIAFGHABxBgwE8AVHpxIAaeAHJ2Y0gG4AUIlTkA3gA8aAFgBMAXzrSgA)

The primary location should be at the function call. Otherwise when we suppress errors, we may add a suppression to the definition of `fn` which could suppress other, legitimate, errors.

In D5989737 and D6253791 I added the ability for errors which somehow point outside of the root location to set the true primary location. (In a style similar to the one added by `Op`.)

In D6055577 and D6238305 I add a `use_op` slot to `GetPropT` and `CallT`.

Together, these changes enable this diff which updates a `use_op` slot on `TypeDestructorT` when we specialize for a function call.

Reviewed By: avikchaudhuri

Differential Revision: D6055682

fbshipit-source-id: 7e36e965e98b3f633f6a361593d7472e1fecb0af