type printer for MatchingPropT
Summary:
refactors the type printer so that we filter out `MatchingPropT`. `type-at-pos` was not impacted, but calling `dump-types` in a file with a sentinel prop test would cause a fatal due to the unhandled `MatchingPropT`. but a `MatchingPropT` is just some internal nonsense we create to defer execution, so it shouldn't ever be printed in the type.

so, this diff makes `type_printer_impl` return an option where `None` means it shouldn't be printed, and makes unions filter them out.

Reviewed By: samwgoldman

Differential Revision: D6390967

fbshipit-source-id: 6ea2e998641c2095304ceb3798548a35179ac1a8