Add mod_op_of_use_t function
Summary:
Adds a `mod_op_of_use_t` function which operates similarly to the `mod_reason_of_use_t` function.

We immeadiately put this function to use in the `FlowConstraint` cache. In future diffs we will also use this function for determining which `use_op` wins when flowing across tvars.

Reviewed By: avikchaudhuri

Differential Revision: D6288979

fbshipit-source-id: 853bbb6fd45d522fb941e5976e7c9903a8353127