Prefer use_ops with non-UnknownUse roots when flowing use_ops through tvars
Summary:
Based on mroch's work in D5899269.

In flows involving type variables like:

```
T ~> UseT(op1, X)
X ~> UseT(op2, U)
```

We have the question of which op to use when `T ~> U`. Do we use `op1`? Do we use `op2`? Previously, we always used `op2`. However, this is often not the right `use_op`.

Consider some use with a `tin` that uses `rec_flow_t`. In that case an `UnknownUse` clobbers the right `use_op`.

In mroch's diff, when we see `T ~> UseT(op1, X)`, we take `op1` and compares it to the `use_op` of a potential `UseT` added to `X` to pick the `use_op` which is not exactly `UnknownUse`. This diff expands on this idea.

In this `use_op` comparison, now called `flow_use_op`, we pick the `use_op` which does not have a _root_ of `UnknownUse`. This means we iterate through `PropertyCompatibility` or `TypeArgCompatibility` `use_op` frames to find the root. If we consider our, common, `tin`/`rec_flow_t` example, we may have started subtyping an object before reaching a type variable.

This diff also performs the comparison for all use types with a `use_op` using the `mod_op_of_use_t` function.

Also, I used this same logic for `X ~> UseT(op2, U)` flows. We accomplish this by tracking the `use_op` which was used when `X` got a lower bound. Then, we take that `use_op` and run it through `flow_use_op` with the new use.

This also enables us to compare all three of the possible `use_op` chains in a flow like `X ~> UseT(op, Y)`. (Where `Y` is also a type variable.) In this case we would see:

```
T ~> UseT(op1, X)
X ~> UseT(op2, Y)
Y ~> UseT(op3, U)
```

...and we need to pick either `op1`, `op2`, or `op3`.

In D5899269, mroch mentioned that it would be useful to have some kind of "join" mechanism for `use_op`s. This would still be useful, however, it is still unclear what such a join mechanism would look like. It is fairly clear, though, that a join operation would almost always entirely pick the `use_op` which does not have a root of `UnknownUse`.

I would stack this diff on top of D5899269, but at this point it's a little old...I'm happy to stack on top of that diff if it lands.

Reviewed By: avikchaudhuri

Differential Revision: D6289433

fbshipit-source-id: 7f34d10ec051496d7fef08444796eb00f4edce3c