Improve React error message use_ops and reason descriptions
Summary:
Improve React error messages by extending the `ReactCreateElementCall` `use_op` with new locations and adding a `ReactGetIntrinsic` `use_op`. In addition to simplifying some reason descriptions.

This needs to be pretty high up in the stack because it depends on some of the hacks around `RReactSFC` being killed in D6290390.

Reviewed By: avikchaudhuri

Differential Revision: D6353419

fbshipit-source-id: a23d8f9683ef71a8a8b84159cb6c087b62df8514