Kill Ops
Summary:
Many of the diffs in this stack replace the need for `Ops` with equivalent, and often better, `use_op`s. At this point, we no longer need `Ops`. Yay\!

Some of the test suite changes may look like regressions. I'm willing to argue that all of the changes are appropriate.

There is one change I'm slightly upset about. In the `react` test suite, some HOC tests lost a pretty valuable `use_op`. This is consistent with all the other errors in that suite, which all point to definitions instead of uses. This issue is fundamentally the same as the issue solved by D6290390 except for type destructors which contain type variables. There are two colliding `use_op`s in type variable flows and we pick the wrong `use_op`.

The fix is likely to expand the implicit type parameter detection mechanism added in D6290390 which wraps relevant `use_op`s in `ImplicitTypeParam`. I have not done this yet since there's no great clean way to do this since we need the implicit instantiation location. This style of React HOC is also an anti-pattern which helps me live with this change.

Reviewed By: avikchaudhuri

Differential Revision: D6353424

fbshipit-source-id: 129f137f1a59a8b36489dbccac080613a17d7613