Dedupe errors on the reverse side of unification
Summary:
Programs like this:

```
declare var any: any;
type X<T> = mixed;

((any: {p: number}): {p: string});
((any: X<number>): X<string>);
```

[try-Flow](https://flow.org/try/#0CYUwxgNghgTiAEA3W8oDsCeAuVmDcAUAC4YAOCAGgDwAqAfPALzwC2AlgB4jCEEAUfdNngBvUjjQBXFgCMQMAL4BKHGJwBnIjDZoA5ssIChOalNny6K+NU3a9lwkA)

Show some redundant errors. This is because when we see `{p: number} ~> {p: string}`, we check both `number ~> string` and `string ~> number`. However, these redundant errors are confusing and frustrating for users. Confusing because the user thinks there are two _seperate_ errors which need to be fixed when in fact there is only one error. (`string` is not compatible with `number`.) Frustrating because each error (especially in the case of unions/intersections) can be 20 lines long. Two identical 20-line-long error messages are not ideal.

This diff dedupes these redundant errors.

How do we do this? By flipping the reasons in `use_op`s for the reverse side of unification.

While in theory, unification is just the flows of `T ~> U` and `U ~> T`. In practice Flow implements a special "optimized" unification algorithm which we describe in notation as `T = U`. This unification algorithm is what we use in most of the areas where we need `T ~> U` and `U ~> T`.

In the unification `T = U`, the Flow `U ~> T` is what I will call the "reverse side of unification." For this side, we introduce a `UnifyFlip` frame `use_op`. So if we were in a `PropertyCompatibility(FunParam(FunCall))` `use_op` chain, the `U ~> T` Flow would make the chain `UnifyFlip(PropertyCompatibility(FunParam(FunCall)))`.

Say we then unified `number = string`. This would produce two incompatibility errrors. One with `UnifyFlip` and one without. In order to deduplicate these error messages we _flip_ the reasons in the `use_op` chain and for the final incompatibility. In `PropertyCompatibility {lower; upper}` we would perform the operation to switch `lower` and `upper`. e.g. `PropertyCompatibility {lower = upper; upper = lower}`.

After flipping, we end up with a `use_op` chain identical to our `use_op` chain without the `UnifyFlip` frame. These two `use_op` chains will produce the same error so our users will only see a single error.

This is all fine and great when we have `number = string` when both `number ~> string` and `string ~> number` are false. In this case directionality does not matter, so "transforming" the second error message to be identical to the first is ok. However, things get a little messier when we have a directional flow like `number = mixed`.

`number ~> mixed` is a valid subtype. However, `mixed ~> number` is not\! If we flip the reasons in our `PropertyCompatibility` error we risk calling things `lower` which are in fact the `upper`. After this diff, `lower` and `upper` might not be the right name for `use_op` reasons. A more accurate name would be `left` and `right` which represent the fact that these reasons no longer have any semantic meaning.

The second concern in flipping the reasons for `number = mixed` is in the final error message. Before we would write "`mixed` is incompatible with `number`." Now, in some cases, we will report "`number` is incompatible with `mixed`." This is fine. The language we use in error messages is not directional. "is incompatible with" is appropriate language for either incompatibility direction.

====================

After this diff, we see a pretty big improvement in the test error messages. This diff removes a lot of unnecessary duplication. For the errors which are directional, we flip the order of some error messages, but this is fine given that "is incompatible with" implies no direction.

This diff is great for our users. I've been asked multiple times about if we can dedupe the two error messages for the same issue. I'm not concerned about the message flippage in some uncommon cases impacting our ability to debug issues. We always have verbose output. If I'm at the point where we I'm analyzing non-symmetric subtyping relations I usually want verbose output anyway. The flippage for `use_op`s is purely asthetic and does not change verbose output in anyway.

Reviewed By: avikchaudhuri

Differential Revision: D6326024

fbshipit-source-id: 65fea4aabc4855969f49ccebc3fd56852920811a